+++ src/collective/blicca/footerblocks/tests/conftest.py
@@ -2,6 +2,7 @@
 
 from pytest_plone import fixtures_factory
 
+from collective.blicca.footerblocks.testing import ACCEPTANCE_TESTING
 from collective.blicca.footerblocks.testing import FUNCTIONAL_TESTING
 from collective.blicca.footerblocks.testing import INTEGRATION_TESTING
 
@@ -10,5 +11,6 @@ globals().update(
     fixtures_factory((
         (INTEGRATION_TESTING, "integration"),
         (FUNCTIONAL_TESTING, "functional"),
+        (ACCEPTANCE_TESTING, "acceptance"),
     ))
 )